Migrated Chat Display into ChatService
Migrated as a public static class.
